instagram-clone/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
python-multipart==0.0.20
Jinja2==3.1.5
requests==2.32.3
python-dotenv==1.0.0
firebase-admin==6.2.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
